refactor: move immutable_string_validator to shared directory
